% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_csv_files_export_results.R
\name{check_csv_folder_export}
\alias{check_csv_folder_export}
\title{Main function to call all csv file checks.}
\usage{
check_csv_folder_export(uploaded_files, wacky_chars_check = FALSE)
}
\arguments{
\item{uploaded_files}{a data frame of data files to be checked with 2 columns:
"name" (containing file name and extension) and "datapath" (containing full file path)}

\item{wacky_chars_check}{a boolean TRUE/FALSE
#' @return A list with two elements:
\describe{
  \item{`result_df`}{A dataframe of file check results}
  \item{`RAC_passed`}{Indicates whether or not all file checks have been passed}
}}
}
\description{
Main function to call all csv file checks.
}
