% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{excel_col_letter}
\alias{excel_col_letter}
\title{Convert numbers to excel-style column letter}
\usage{
excel_col_letter(col_index)
}
\arguments{
\item{col_index}{A positive integer representing the column index.}
}
\value{
A character string representing the Excel column letter(s)
}
\description{
Converts a numeric column index (e.g., 1, 2, ..., 28) into the corresponding Excel-style column
letters (e.g., "A", "B", ..., "AB").
}
\keyword{internal}
