% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headers.R
\name{identify_spaces_in_headers}
\alias{identify_spaces_in_headers}
\title{Identify spaces in the headers of a Data Frame}
\usage{
identify_spaces_in_headers(raw_content_df)
}
\arguments{
\item{raw_content_df}{A data frame containing raw content, from the read_csv_raw function.}
}
\value{
A single character string listing all flagged cells, or an empty string if none found.
}
\description{
This function scans all column names in a data frame and identifies whitespace characters.
 It flags and returns a string of column headers with their spreadsheet-style cell references (e.g. A: Air temp, B: Mass (kg)).
}
