% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_fun.R
\name{random_spinner}
\alias{random_spinner}
\title{Choose and display a random spinner from the waiter package}
\usage{
random_spinner()
}
\value{
Called for its side effect (displays a spinner). Returns `NULL` invisibly.
}
\description{
Helper function to choose a random spinner to show while code is executing. No input arguments.
}
\keyword{internal}
