% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_raw.R
\name{read_csv_raw}
\alias{read_csv_raw}
\title{Read in data from csv safely to detect encoding issues}
\usage{
read_csv_raw(file_path)
}
\arguments{
\item{file_path}{The file path to a csv file}
}
\value{
A list with two elements:
\describe{
  \item{`raw_content`}{A single UTF-8 character string of the sanitized file contents.}
  \item{`raw_content_df`}{A data frame parsed from the sanitized content.}
}
}
\description{
This function optimistically reads in a file assuming utf-8 encoding. Forces any non UTF-8
characters into valid UTF-8, then detects any replaces or hex escapes that were made.
If made, the file is read in as raw-bytes and the encoding estimated.
The BOM is stripped if present. The content is parsed to a data frame and to a single UTF-8 character string.
}
